// src/core/types/logger/logger.ts
export interface Logger {
  log: (message?: any, ...optionalParams: any[]) => void;
  error: (message?: any, ...optionalParams: any[]) => void;
  info: (message?: any, ...optionalParams: any[]) => void;
  warn: (message?: any, ...optionalParams: any[]) => void;
}

export interface LoggerOptions {
  enable: boolean;
  logger?: Logger;
}